+++ ChangeLog.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ## Release 1.7
+- NEW: ST-7 Add user attachments when generating supplier RFQs (upload, store on proposal, include in emails, remove before confirm) - *15/12/2025* - 1.7.2
 - FIX : When converting a quote into a price request, VAT must be charged. - *05/11/2025* -1.7.1
 - NEW : Add mass action to create a new supplier proposal from a customer proposal or order - *09/10/2025* - 1.7.0